Temporarily hardcode ~tildes.official as the main user count.
With the new groups being added some of them have more subscribers
than ~tildes.official which is making the homepage chart only have
3 datapoints (as of today).

I want to revert get_highest_subscribers back but add a check so
it gets the group with the highest subscribers and a certain amount
of rows. But I have yet to figure out how to do that.